* 🎨 style(Experience.tsx): update Experience component's detail list style
* 🎨 style(Resume.tsx): remove ContactButton component and replace it with a Link component for email address
* 🎨 style(Resume.tsx): adjust max-width of container to 750px

* 🐛 fix(NavMenu.tsx): fix pathname for blog posts
* ✨ feat(NavMenu.tsx): add toggle button for light/dark mode

* ♻️ refactor(NavMenu.tsx): simplify NavMenu component by using LayoutGroup component
* ✨ feat(NavMenu.tsx): add dynamic rendering of navigation items using Object.entries and map function

* ✨ feat(Contact.tsx): add Contact component with form to send email message

* ✨ feat(contact): add Contact component and ContactPage
* 🎨 style(layout.tsx): remove ActionIcon from Layout component

* ✨ feat(Footer.tsx): add social media icons and links to the footer

* 🗑️ chore(ActionIcon.tsx): delete ActionIcon component file

* 🗑️ chore(Contact.tsx): delete Contact component file

* 🗑️ chore(ContactButton.tsx, index.ts): remove ContactButton component
* 🗑️ chore(index.ts): remove ContactButton and Contact components from ActionIcon export
