+++ src/app/(main)/contact/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Contact } from "./Contact";
