+++ src/app/(main)/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ActionIcon, Page } from "@chia/ui";
+import { Page } from "@chia/ui";
 import { NavMenu, ReduxProvider } from "./components";
 import type { ReactNode } from "react";
 
@@ -8,7 +8,6 @@ const Layout = ({ children }: { children: ReactNode }) => {
   return (
     <ReduxProvider>
       <NavMenu />
-      <ActionIcon />
       <Page>{children}</Page>
     </ReduxProvider>
   );
