--- src/ui/ActionIcon/index.ts
@@ -1,3 +1,0 @@
-export { default } from "./ActionIcon";
-export { default as ContactButton } from "./ContactButton";
-export { default as Contact } from "./Contact";
